requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.9,<3.0
